Navigator 위치 변경
