+++ src/Navigator/Navigator.tsx
@@ -1,16 +1,16 @@
 import * as React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import {StackParamsList} from './types/types';
-import useStore from './Store/store';
+import {StackParamsList} from '../types';
+import useStore from '../Store/store';
 
 // Splash
-import {Splash} from './Screens/Splash';
+import {Splash} from '../Screens/Splash';
 
 // 메인 서비스 스크린
-import {Home} from './Screens/Main/Home';
+import {Home} from '../Screens/Main/Home';
 
 // 인증 관련 스크린
-import {Auth} from './Screens/Auth/Auth';
+import {Auth} from '../Screens/Auth/Auth';
 
 const Stack = createNativeStackNavigator<StackParamsList>();
