+++ src/types.ts
@@ -1,7 +1,13 @@
-export type MainStackParamList = {
-  HomeScreen: undefined;
-};
+export type StackParamsList = {
+  // Splash
+  Splash: undefined;
 
-export type AuthStackParamList = {
-  AuthScreen: undefined;
+  // 메인 서비스 스택
+  Home: undefined;
+
+  // 인증 관련 스택
+  Auth: {
+    setIsLoggedIn: React.Dispatch<React.SetStateAction<boolean>>;
+    a: string | undefined;
+  };
 };
